Adding complex transcendental functions. Fixing few bugs (static).

+++ include/Core/Traits.hpp
@@ -125,7 +125,7 @@ namespace Kartet
 	struct Traits<long double>
 	{
 		typedef void SubTraits;
-		typedef float BaseType;
+		typedef long double BaseType;
 		
 		static const bool 	isConst 	= false,
 					isPointer 	= false,
